worker/requirements.txt
celery==5.3.4
redis==5.0.1
monai==1.4.0
nibabel==5.1.0
numpy>=1.22.0,<2.0.0
einops==0.7.0
antspyx==0.6.3
pydicom
dicom2nifti